Compras Finalizadas

+++ src/main/java/co/edu/unbosque/ProyectoBases/MariaDB/model/Compra.java
@@ -13,49 +13,60 @@ public class Compra {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-	
+
 	@ManyToOne
 	private Cliente idCliente;
 	private double precio;
 	private LocalDate fecha;
-	private Integer idCasa;
-	private Integer idApartamento;
+	private Long idCasa;
+	private Long idApartamento;
+
 	public Long getId() {
 		return id;
 	}
+
 	public void setId(Long id) {
 		this.id = id;
 	}
+
 	public Cliente getIdCliente() {
 		return idCliente;
 	}
+
 	public void setIdCliente(Cliente idCliente) {
 		this.idCliente = idCliente;
 	}
+
 	public double getPrecio() {
 		return precio;
 	}
+
 	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
+
 	public LocalDate getFecha() {
 		return fecha;
 	}
+
 	public void setFecha(LocalDate fecha) {
 		this.fecha = fecha;
 	}
-	public Integer getIdCasa() {
+
+	public Long getIdCasa() {
 		return idCasa;
 	}
-	public void setIdCasa(Integer idCasa) {
+
+	public void setIdCasa(Long idCasa) {
 		this.idCasa = idCasa;
 	}
-	public Integer getIdApartamento() {
+
+	public Long getIdApartamento() {
 		return idApartamento;
 	}
-	public void setIdApartamento(Integer idApartamento) {
+
+	public void setIdApartamento(Long idApartamento) {
 		this.idApartamento = idApartamento;
 	}
-	
-	
+
 }
